+++ website/src/main/java/com/acme/homeloans/model/Submission.java
@@ -6,9 +6,9 @@ package com.acme.homeloans.model;
 public class Submission {
 
     private String email;
-    private float price;
-    private float loanAmount;
-    private float salary;
+    private double price;
+    private double loanAmount;
+    private double salary;
 
     private int creditScore;
 
@@ -20,27 +20,27 @@ public class Submission {
         this.email = email;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 
-    public float getLoanAmount() {
+    public double getLoanAmount() {
         return loanAmount;
     }
 
-    public void setLoanAmount(float loanAmount) {
+    public void setLoanAmount(double loanAmount) {
         this.loanAmount = loanAmount;
     }
 
-    public float getSalary() {
+    public double getSalary() {
         return salary;
     }
 
-    public void setSalary(float salary) {
+    public void setSalary(double salary) {
         this.salary = salary;
     }
